Making things work with the Huzzah32.
For the moment, I've reassigned pins in the firmware. I am not 100%
certain this was necessary. However, I am confident of the mapping for
SPI, as the SPI pins are routed differently on an AirLift than they are
on a Huzzah.

That's kinda annoying. Why change the pins from one WROOM to the next?
There might be a pun in there.

With a pin change, I get communication between a Feather M4 running
CircuitPython (the test harness) and the Huzzah32.