Adding in I2CS class.
This is a first pass at the I2C class. It is almost 100% certain not to
work... because I haven't implemented anything yet.

That said, it compiles, which is also a first step. In theory, an
I2C object is created, and it has a .transfer() method, which is what
I need to work on further.

However, there's not a lot on the firmware side. Once I'm
sending/receiving commands via I2C, it's largely on the client side that
some changes have to happen.

Touch wood.