Works, but unhappy.
I have I2C traffic moving back-and-forth between a Feather M4 (running
CircuitPython) and the ESP32 on a Huzzah32 board. The CP side is not the
full Adafruit ESP32 library: it is just me sending I2C messages using the raw
I2C library.

That said, I can invoke the pinMode() and digitalWrite() commands
(command numbers 0x50 and 0x51, respectively) in the NINA firmware.

BUT, a big caveat: there seem to be some uglies hiding in the FreeRTOS
layer. This means that reading in multiple bytes with I2C (a la the SPI
version) is going to be a potential no-go.

The fix to this is that I think I need to write a byte-by-byte state machine for the I2C version.

SOME DETAILS:

Reading this post from May 2019:
https://esp32.com/viewtopic.php?t=10758

It looks like there is a possible interaction happening between the
process of reading the I2C port and the underlying FreeRTOS ring
buffer... as in, the ring buffer might not get cleared. And, calling

i2c_reset_rx_fifo(_i2cPortNum);

might not make a difference. (It doesn't in my code.)

That said, if I do the same thing I did on the "I2C OpenLog":

https://bitbucket.org/corounum/i2c-logger-firmware/src/5045da9410a48c9de485a104333a7b5403df476a/firmware/firmware.ino#lines-150

I would split things here in this implementation, but at the end of the
day, I would read the bytes one at a time.

* First, a header.
* Second, an indicator as to how many parameters you're sending.
* Third, the length (in bytes) of each of those params.
  (This suggests that a parameter maxes out at 255 bytes in length. This
   can be managed, either by allowing all parameters to be 16 bits, or
   longer, or...)
* Fourth, each parameter.
* Finally, a CRC.

This allows the I2C device to read in an amount of data specified by the
client, as opposed to attempting to fill a buffer, or wait for an I2C
stop event. The client needs to know how much data it is sending... BUT,
the client does not have to buffer it all. (For example, an ATMega328
could say it is going to send 1,000,000 bytes... and do so by sending
0xDECAFBAD over and over.)

On the receing side, we can set limits... for eaxmple, we could say
"inputs cannot exceed [insert number of bytes here]."

The command handler (CommandHandler.cpp) shouldn't have to change. The
sketch (the setup() and loop(), if you will) will have some big
differences from the Arduino/SPI version if I go down this route.