Moving towards interrupt-driven code.
This is not really interrupt-driven.

This spawns a second task, which runs in parallel to the main loop.

That loop constantly checks to see if there is a byte waiting on the I2C
bus. If there is, it will (eventually) stuff that in a buffer.

When that buffer contains a complete command (as per a state machine),
it will signal to the main loop that it can process that command.

At that point, there is the question of an I2C response. That may, or
may not, become another task.

If these are separate tasks, then semaphores or some other concurrency
control mechanism will be needed. Some thought will be needed there, as
there are possibilities of bad behavior on the client-side. What if
someone makes multiple I2C calls that "expect" responses, but they don't
read the response... should that be possible? Can I (for example) fetch
a web page but not read the response from the web server?

There are lower-level routines (_isr() routines) that need to be
investigated. This would move this from a polling driver to an
interrupt-driven driver. The fundamental structure---of watching for
bytes, running a state machine, and then processing a buffer when the
state machine is successful---would remain the same.

I have to investigate why, depending on CP I2C choices, I get different
behavior in my "debug" printing. On one hand, I get bytes "as they come
in," and on another, I get them... buffered. Why would flushing of the
UART be impacted by my choice of library. I'm wondering about STOP
conditions and other things, but my code doesn't suggest that should
matter.

Things to investigate.