Working with I2C from CircuitPython.
@ladyada @ptorrone

This is another step towards the working state machine exchange in CP.

Note that the repository

https://github.com/jadudm/nina-fw-testing

has some scripts that will help test this firmware.

I'm currently able to send commands via I2C---in particular, I'm testing
the pinMode() and digitalWrite() commands. At the core, a command on the
client side might look like:

def digitalWrite(pin, lvl):
  ls = [[_SET_DIGITAL_WRITE_CMD], [pin], [lvl]]
  write_msg("dW", ls)

A programmer on (say) a Feather M4 connected to a Huzzah32 running the
firmware might then write

digitalWrite(13, HIGH)

(with HIGH == 1) and the result would be the sending of an I2C message
to the Huzzah.

This involves calling:

def write_msg(tag, ls, addr = 0x2A, stp = False):
  # print(tag + "\n")
  # End byte is exclusive; add one.
  end = make_msg(ls) + 1
  with device:
    device.write(bytes(msg), end = end)

def make_msg(ls):
  msg[0] = 0xAD
  msg[1] = 0xAF
  msg[2] = len(ls)
  c1 = 0
  for subls in ls:
    msg[3 + c1] = len(subls)
    c1 = c1 + 1
  c2 = 0
  for subls in ls:
    for e in subls:
      msg[3 + c1 + c2] = e
      c2 = c2 + 1
  # NON CRC
  msg[3 + c1 + c2] = 255
  return (3 + c1 + c2)

which does not yet correclty calculate a CRC. That aside, it assembles a
command, which involves:

A header (two bytes)
The number of values being sent (one byte)
The length of each value (one byte each, so in this case, three bytes
for three values)
The values themselves (again, three bytes)
A CRC (not yet computed, so stuffed in as 0xFF)

This packet encodes the command, the pin number, and the level (HIGH or
LOW). The protocol allows for a variable number of parameters of
variable length.

The length returned is to make sure we only send the portion of the
buffer encoding the command.

We now need to implement the read side of the protocol. This must be
either manually carried out by the client, or wrapped in the sending
abstraction. For many commands (eg. pinMode(), digitalWrite()), it is
unclear whether you would want to call the read. On the device, we'll
buffer the response, but it might never be retreived.

After implementing the read side of the equation (which shouldn't take
long... we were slowed down by the 4th and travel to Germany for a
conference), it really is a matter of implementing the client library. I
don't think it is going to look the same (per say... the API might be
the same) as the SPI library. The concerns in terms of
bi-directionality, etc., of a SPI library are rather different than the
send/receive nature of an I2C library.