Better handling of I2C messages.
@ladyada

Now deciding that I2C messages "just work."

After doing more reading/thinking, I've decided to just decide that I2C
messages will "just work." This seems reasonable; how many other things
are going on on the board that cannot be checked/backed off/retried?
Further, there's no way to implement an error handling protocol in (say)
Makecode.

As a first step, this means that I just read in the right number of
bytes. If the client says they have 3 parameters, and the first one is
1 byte, the next is 32 bytes, and so on... I read that many bytes.

I may have to think about some timeouts. But, the initial test, with a
modified SPI library (that is the Adafruit library modulo sending I2C
messages instead of SPI) works.

Well, it sets a pinmode and blinks pin 13.

But, this means that I haven't replaced the library. As a first step,
makes clear what needs to be refactored.

Next steps are to think a bit more about timewouts in the firmware
(which will be simplified by thinking about the ready pin), implementing
return values, and... refactor for I2C.

This is better than all the work that I did before, which was very
complex.